fix(sql_server): each instances is a different software
Do not treat one single server with multiples sql server instances as a
single entity.
Each sql server should return a different software entry.

The error was to get the InstalledInstances for each sql server.

The output of the "extra_data" for sql servers is changed:
 - added a new key "instance" with the instance name.
 - removed the "databases_per_instance" dict ({instance: [db1,db2]})
 - added "databases" list

Added a new unit test with two sql server instances.